+++ docs/versionhistory.rst
@@ -11,6 +11,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - **BACKWARD INCOMPATIBLE** Dropped the ``resolve_reference()`` utility function
 - **BACKWARD INCOMPATIBLE** Dropped the ``TeardownError`` exception in favor of PEP 654
   exception groups
+- **BACKWARD INCOMPATIBLE** Dropped the deprecated ability to use a ``Context`` as a
+  synchronous context manager
 - **BACKWARD INCOMPATIBLE** Dropped the ``--unsafe`` switch for ``asphalt run`` –
   configuration files are now always parsed in unsafe mode
   (use the ``!!python/name`` YAML tag instead and replace ``:`` with ``.`` in the